added spawnpoint where enemies will spawn from (still by pressing E button). enemy will spawn from a random spawnpoint that is offscreen

+++ genericwave/Bullet.cpp
@@ -26,7 +26,7 @@ void Bullet::move() {
 
 }
 
-void Bullet::setPosition(int x, int y) {
+void Bullet::setPosition(double x, double y) {
     this->x = x;
     this->y = y;
     this->sprite.setPosition(x, y);
